调整小程序订阅消息接口类和返回对象包路径
